requirements.txt
flask
flask-cors
pandas
openpyxl
plotly
pyyaml
requests
azure-mgmt-compute==30.6.0
azure-mgmt-costmanagement==4.0.1
azure-identity==1.17.1
azure-mgmt-resource==23.1.1
gitpython